requirements.txt
absl-py==1.1.0
alembic==1.7.7
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.2.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.1.5
attrs==21.4.0
autopage==0.5.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
blis==0.7.8
blosc==1.7.0
boto3==1.26.115
botocore==1.29.115
Brlapi==0.7.0
bs4==0.0.1
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.0
chardet==3.0.4
charset-normalizer==3.1.0
chrome-gnome-shell==0.0.0
click==8.1.3
cliff==3.10.1
cloudpickle==1.3.0
cmaes==0.8.2
cmd2==2.4.1
colorama==0.4.5
colorlog==6.6.0
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.6
dask==2.8.1+dfsg
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
dm-tree==0.1.8
duplicity==0.8.12.0
einops==0.4.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.0.1
etils==1.2.0
exceptiongroup==1.1.0
executing==0.8.3
fasteners==0.14.1
ffmpeg-python==0.2.0
filelock==3.8.0
flatbuffers==23.3.3
fonttools==4.33.2
fsspec==0.6.1
future==0.18.2
gast==0.4.0
gensim==4.1.2
google-api-core==2.11.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-cloud-vision==3.4.1
google-pasta==0.2.0
google-search-results==2.4.1
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==1.1.2
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
h5py==3.7.0
html5lib==1.0.1
httplib2==0.14.0
icecream==2.1.3
idna==2.8
imageio==2.4.1
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jax==0.4.8
jdcal==1.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.1.0
jsonpickle==3.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.12.0
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
keyring==18.0.1
kiwisolver==1.4.2
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.1
lightgbm==3.3.2
locket==0.2.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
ml-dtypes==0.1.0
monotonic==1.5
more-itertools==4.2.0
murmurhash==1.0.8
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.10.4
networkx==2.8.5
nltk==3.7
notebook==6.4.10
numexpr==2.7.1
numpy==1.23.0
nvidia-ml-py==11.495.46
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
openpyxl==3.0.3
opt-einsum==3.3.0
optuna==2.10.0
outcome==1.2.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
partd==1.0.0
pathy==0.6.2
patsy==0.5.2
pbr==5.8.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.2
plotly==5.7.0
pluggy==0.13.0
preshed==3.0.7
prettytable==3.2.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.8.1
py4j==0.10.9.3
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pybase64==1.2.3
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
pydantic==1.9.2
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyperclip==1.8.2
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pyspark==3.2.1
pytesseract==0.3.10
pytest==4.6.9
pytextrank==3.2.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-digitalocean==1.17.0
python-dotenv==0.21.1
python-vlc==3.0.16120
pytz==2022.1
PyWavelets==0.5.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.3.15
reportlab==3.5.34
requests==2.28.2
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
s3transfer==0.6.0
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
SecretStorage==2.3.1
selenium==4.8.0
Send2Trash==1.8.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.35
srsly==2.4.4
stack-data==0.2.0
statsmodels==0.13.2
stevedore==3.5.0
systemd-python==234
tables==3.6.1
tenacity==8.0.1
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-dataset==0.1.1
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-examples===3df76a9e45e3a1d6165e42c565a2f8622f7c1f8c-
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.13.1
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
thinc==8.1.0
threadpoolctl==3.1.0
tinytag==1.8.1
toml==0.10.2
toolz==0.9.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.4.2
typing_extensions==4.2.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.16.3
wadllib==1.3.3
wasabi==0.10.1
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wordcloud==1.8.1
wrapt==1.14.1
wsproto==1.2.0
xgboost==1.6.0
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
zipp==3.7.0
